finished rest of the page
